Fix tilesheet HTML and simplify inline styling
Changes tilesheet HTML to a span, so the tile can be used inside a link.
This is also more semantically correct, as a tile is not a divider. Also
uses a <br> instead of a &nbsp; for an empty element, as it doesn't
create an extra space.
Moves default tile styling to a stylesheet. This is loaded
unconditionally, rather than only when a tilesheet is used, as
re-downloading 125 KB to save 70 bytes on the few pages which don't have
tilesheets doesn't seem worth it.
Also simplifies the remaining inline styling by: removing the quotes
from the url() as getUrl() should return a properly escaped url, and
merging the position with the background shorthand.