Tiles can now have their display names and descriptions translated
* Add a new parser function #iconloc which takes the item and mod names, the type, and the language, and returns the localized data for that type (name or description) and language.
* Add a new special page TileTranslator for translating existing translations, and creating new ones. A link to this page for each entry is added to the TileList table, just like the Edit link.
* Fix a couple issues with CreateTileSheet that are somewhat unrelated, but prevented me from testing.

Add an entirely new table, ext_tilesheet_languages, which contains the columns entry_id (not a primary key), lang, display_name, and description. entry_id is identical to that in ext_tilesheet_items. lang contains language codes, and display_name and description are self explanatory. entry_id is not a primary key because we need it to be able to be duplicated for multiple different langauges.
The SQL install code will populate the language table with all of the existing item entries, defaulting the language to en and the name to the item_name.

Close #8